requirements.txt
Flask
PyPDF2
pdfplumber
Pillow
PyMuPDF
